requirements.txt
boto3
matplotlib
cartopy
osmnx
arm_pyart
